requirements.txt
annotated-types==0.7.0
anyio==4.9.0
cachetools==5.5.2
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
cssselect==1.3.0
emoji==1.6.3
facebook-sdk==3.1.0
google-auth==2.40.3
google-genai==1.24.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jmespath==1.0.1
lxml==6.0.0
markdown-it-py==3.0.0
mdurl==0.1.2
packaging==25.0
parsel==1.10.0
pillow==11.3.0
pilmoji==2.0.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
python-dotenv==1.1.1
requests==2.32.4
rich==14.0.0
rsa==4.9.1
schedule==1.2.2
shellingham==1.5.4
sniffio==1.3.1
tenacity==8.5.0
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
w3lib==2.3.1
websockets==15.0.1
